moved to header only
* this makes it compatible with existing code
* maps are now defined statically
